Add Dockerfile for League-Prod-Toolkit
Dockerfile support for the Web App.

I do not have experience with setting up Github actions, but I would like to set that up for building the Image automatically on new releases.

The Docker Image contains all the modules installed by default. I can understand that this is probably not the smartest, but from testing it was the best as it allowed for quicker initial setup, and less problems.

Docker Image avaliable on my personal account. This is something I would prefer by moved to an "official" RCVolus docker account: https://hub.docker.com/r/eirikskarding/leagueprodtoolkit